+++ substra/sdk/__init__.py
@@ -1,7 +1,7 @@
 from substra.sdk import models
 from substra.sdk import schemas
 from substra.sdk.client import Client
-from substra.sdk.config import BackendType
+from substra.sdk.schemas import BackendType
 from substra.sdk.utils import retry_on_exception
 
 __all__ = [